@@ -6,7 +6,7 @@ import axios from 'axios'
 
 import App from './App.vue'
 import router from './router'
-import VueTailwindDatepicker from "vue-tailwind-datepicker"
+import VueTailwindDatepicker from 'vue-tailwind-datepicker'
 
 axios.defaults.withCredentials = true
 
@@ -14,7 +14,6 @@ const app = createApp(App)
 
 app.use(createPinia())
 app.use(router)
-app.use(VueTailwindDatepicker);
+app.use(VueTailwindDatepicker)
 
 app.mount('#app')
-
